Allow pickling simple instances of caches.
This is not particularly robust, but it handles (and tests) the simplest
examples.  In general, pickle does not handle lambda functions, so this
is about the best that can be done.  It's worth noting that this uses
a non-default pickle protocol: the regular one fails due to __slots__
with no __getstate__ implementation.  This non-default protocol is
specified with dumps(cache, -1): the -1 means "latest".

For a more robust pickling solution, try the dill module.