func, decorators: prefer values already in the cache in races
This change updates the behavior of the cache decorators to prefer to
return the value *already in the cache* when the backing function is
called simultaneously from two separate threads. Previously, if two
threads raced to the cache, each would get the version of the cached
value calculated by its own thread, and the first's value would be
immediately evicted by the second.

In cases where caches are used to manage expensive resources, this could
result in two separate expensive objects hanging around the system
at the same time. With this change, there is only one long-lived object
exposed to callers, and the transient second object created by the race
can be immediately collected by the runtime.